api/message/requirements.txt
# Core
python-dotenv>=1.0.1
openai>=1.25.0
pydantic>=2.7.1

# LangChain ecosystem
langchain-core>=0.1.47
langchain-openai>=0.1.6
langchain-azure-ai>=0.1.5
langgraph>=0.0.40

# Vector database
pinecone-client[grpc]>=4.0.0

# Azure AI SDK
azure-ai-inference>=1.0.0b1
azure-core>=1.29.5

# Web back‑end
fastapi>=0.111.0
uvicorn[standard]>=0.28.0

# Images / files
pillow>=10.3.0